+++ lib/hitnmiss/driver.rb
@@ -1,12 +1,14 @@
 module Hitnmiss
   module Driver
     class Hit
-      attr_reader :value, :updated_at, :fingerprint
+      attr_reader :value, :updated_at, :fingerprint, :last_modified
 
-      def initialize(value, updated_at: nil, fingerprint: nil)
+      def initialize(value, updated_at: nil, fingerprint: nil,
+                     last_modified: nil)
         @value = value
         @updated_at = updated_at
         @fingerprint = fingerprint
+        @last_modified = last_modified
       end
     end