+++ lib/hitnmiss/entity.rb
@@ -1,11 +1,12 @@
 module Hitnmiss
   class Entity
-    attr_reader :value, :expiration, :fingerprint
+    attr_reader :value, :expiration, :fingerprint, :last_modified
 
-    def initialize(value, expiration_in_seconds=nil, fingerprint=nil)
+    def initialize(value, expiration: nil, fingerprint: nil, last_modified: nil)
       @value = value
-      @expiration = expiration_in_seconds
+      @expiration = expiration
       @fingerprint = fingerprint
+      @last_modified = last_modified
     end
   end
 end